+++ src/LyraNetwork/Constants.php
@@ -4,5 +4,4 @@ namespace LyraNetwork;
 class Constants
 {
     const SDK_VERSION = '3.0.2';
-    const END_POINT   = 'https://secure.payzen.eu/api-payment/V3/';
 }